--- robomae/labels.json
@@ -1,4 +1,0 @@
-{
-"basiclabels":["sit","walk", "stand","lying"],
-"highlevellabels":["eating","watching tv","pick up pill"]
-}
