--- robomae/test/rosbags/AudioBags/2016-07-04-16-19-14_audio.csv
@@ -1,3 +1,0 @@
-867.8254928315412,3520.451612903226,Speech::George
-5670.805331541219,8543.175627240144,Speech::Eleutheria
-11698.07414874552,14476.268369175627,Speech::Maria
