--- robomae/test/rosbags/AudioBags/2016-07-04-16-19-34_audio.csv
@@ -1,3 +1,0 @@
-776.3347771904465,3375.659270428602,Speech::George
-4044.946415394115,6862.178351179182,Speech::Eleutheria
-6877.743168503962,9461.502844417337,Speech::Maria
