--- robomae/test/rosbags/AudioBags/2016-07-22-13-20-07_audio.csv
@@ -1,5 +1,0 @@
-2000.0,20000.0,Music
-21000.0,24000.0,Music
-24000.0,32000.0,Activity
-35000.0,37000.0,Speech::Eleni
-38000.0,41000.0,Laugh
