--- robomae/test/rosbags/AudioBags/2016-07-22-13-24-10_audio.csv
@@ -1,3 +1,0 @@
-1000.0,15000.0,Music
-16000.0,24000.0,Speech::Eleni
-24842.940108302755,27849.851905224823,Activity
