--- robomae/test/rosbags/AudioBags/2016-07-22-13-25-52_audio.csv
@@ -1,6 +1,0 @@
-2000.0,3000.0,Activity
-3000.0,4000.0,Music
-4000.0,5000.0,Speech
-5000.0,10000.0,Activity
-11000.0,13000.0,Speech
-15000.0,20000.0,Activity
