--- robomae/test/rosbags/ss1_lsN_sc1B_rugsta_cg_v_audio.csv
@@ -1,10 +1,0 @@
-15000.0,17000.0,Activity
-35476.54745657569,54000.0,Activity
-56000.0,62000.0,Activity
-64764.286290322576,75000.0,Activity
-8638.328598014887,10448.843362282876,Activity
-18542.90936724566,33346.53008684863,Activity
-62208.26544665012,63805.77847394541,Speech::George
-87129.46867245657,88407.4790942928,Activity
-75414.373138957817,92348.01122828784,TV
-94904.032071960304,95862.539888337473,Activity
