--- robomae/test/rosbags/ss1_lsN_sc3A_rudmal_cggsta_v_audio.csv
@@ -1,19 +1,0 @@
-3819.575930521092,5624.166004962779,Steps
-5624.166004962779,6724.525806451613,Speech::George
-6812.554590570718,8573.130272952852,Speech::Dimitris
-8705.173449131515,9629.475682382132,Speech::George
-10377.720347394539,12050.267245657569,Speech::George
-12050.267245657569,12798.511910669973,Laugh
-12842.526302729528,13326.684615384615,Speech::George
-15747.476178660047,16627.764019851118,Speech::George
-16979.879156327541,18960.526799007443,Speech::Dimitris
-19048.555583126548,19664.757071960295,Speech::George
-19928.843424317616,21029.20322580645,Speech::Dimitris
-21117.232009925559,25826.771960297767,Laugh
-29039.822580645163,30096.167990074438,Speech::George
-30536.311910669978,31416.599751861038,Laugh
-31548.642928039702,32516.95955334988,Speech::George
-32649.002729528533,34409.578411910676,Speech::Dimitris
-34761.693548387098,35421.909429280393,Speech::George
-26090.858312655084,28643.693052109182,Speech::George
-26971.146153846152,27367.275682382133,Speech::Dimitris
